refactoring: move UTM zone class to production module #578
