+++ tests/utm_test.py
@@ -5,10 +5,11 @@ import pytest
 import sqlalchemy
 from django.conf import settings
 from django.contrib.gis.geos import Point
-from sqlalchemy.engine.url import URL as DBURL
 from sqlalchemy import select, func
+from sqlalchemy.engine.url import URL as DBURL
 
 from osmaxx.conversion_api.coordinate_reference_systems import WGS_84
+from osmaxx.geodesy.coordinate_reference_system import UniversalTransverseMercatorZone
 from tests.utils import slow
 
 MAX_LONGITUDE_OFFSET = 90.0 - 9.9e-14
@@ -53,29 +54,6 @@ def transformable_point_latitude_degrees(request):
 @pytest.fixture
 def utm_zone(hemisphere, utm_zone_number):
     return UniversalTransverseMercatorZone(hemisphere=hemisphere, utm_zone_number=utm_zone_number)
-
-
-class UniversalTransverseMercatorZone:
-    HEMISPHERE_PREFIXES = dict(
-        north=326,
-        south=327,
-    )
-    NUMBER_OF_ZONES_PER_HEMISPHERE = 60
-    ZONE_WIDTH_DEGREES = 360 / NUMBER_OF_ZONES_PER_HEMISPHERE
-
-    def __init__(self, hemisphere, utm_zone_number):
-        assert hemisphere in self.HEMISPHERE_PREFIXES.keys()
-        assert utm_zone_number - 1 in range(60)
-        self.hemisphere = hemisphere
-        self.utm_zone_number = utm_zone_number
-
-    @property
-    def srid(self):
-        return self.HEMISPHERE_PREFIXES[self.hemisphere] * 100 + self.utm_zone_number
-
-    @property
-    def central_meridian_longitude_degrees(self):
-        return -180 + (self.utm_zone_number - 0.5) * self.ZONE_WIDTH_DEGREES
 
 
 @pytest.fixture(params=UniversalTransverseMercatorZone.HEMISPHERE_PREFIXES.keys())
